jump to my location button added in request screen
